+++ packages/ui/Tooltip/Tooltip.tsx
@@ -1,25 +1,10 @@
-import type { HTMLAttributes} from 'react';
-import { createContext, useContext } from 'react';
-import TooltipTrigger from 'Tooltip/Trigger';
-import TooltipContent from 'Tooltip/Content';
-import { useTooltip } from 'Tooltip/useTooltip';
+import type { HTMLAttributes } from 'react';
+import TooltipTrigger from './Trigger';
+import TooltipContent from './Content';
+import { useTooltip } from './useTooltip';
 import * as S from './style.css';
 
-export interface TooltipContextProps {
-  isTooltipVisible: boolean;
-  position: 'top' | 'bottom';
-  contentRef: React.RefObject<HTMLDivElement>;
-}
-
-export const TooltipContext = createContext<TooltipContextProps | null>(null);
-
-export const useTooltipContext = () => {
-  const context = useContext(TooltipContext);
-
-  if (!context) throw new Error('Tooltip components must be used within <Tooltip.Root>');
-
-  return context;
-};
+import { TooltipContext } from './TooltipContext';
 
 function TooltipRoot({ children, ...props }: HTMLAttributes<HTMLDivElement>) {
   const { isTooltipVisible, triggerRef, contentRef, position } = useTooltip();
