+++ packages/ui/Tooltip/index.ts
@@ -1,1 +1,3 @@
 export { default as Tooltip } from './Tooltip';
+export { default as TooltipTrigger } from './Trigger';
+export { default as TooltipContent } from './Content';
